+++ weconnect_mqtt/weconnect_mqtt_base.py
@@ -95,6 +95,8 @@ def main():  # noqa: C901  # pylint: disable=too-many-branches,too-many-statemen
     weConnectGroup = parser.add_argument_group('WeConnect')
     weConnectGroup.add_argument('-u', '--username', type=str, help='Username of Volkswagen id', required=False)
     weConnectGroup.add_argument('-p', '--password', type=str, help='Password of Volkswagen id', required=False)
+    weConnectGroup.add_argument('-s', '--spin', help='S-PIN of Volkswagen id, required for selected commands', required=False, nargs='?', action='store',
+                                default=None, const=True)
     defaultNetRc = os.path.join(os.path.expanduser("~"), ".netrc")
     weConnectGroup.add_argument('--netrc', help=f'File in netrc syntax providing login (default: {defaultNetRc}).'
                                 ' Netrc is only used when username and password are not provided  as arguments', default=None, required=False)
@@ -163,6 +165,7 @@ def main():  # noqa: C901  # pylint: disable=too-many-branches,too-many-statemen
 
     username = None
     password = None
+    spin = None
 
     if args.username is not None and args.password is not None:
         username = args.username
@@ -193,6 +196,26 @@ def main():  # noqa: C901  # pylint: disable=too-many-branches,too-many-statemen
                 sys.exit(1)
             username = args.username
             password = getpass.getpass()
+
+    if args.spin is not None:
+        spin = args.spin
+    else:
+        if args.netrc is not None:
+            netRcFilename = args.netrc
+        else:
+            netRcFilename = defaultNetRc
+        try:
+            secrets = netrc.netrc(file=args.netrc)
+            _, account, _ = secrets.authenticators("volkswagen.de")
+            if account is not None:
+                spin = account
+        except netrc.NetrcParseError as err:
+            LOG.error('Authentification using .netrc failed: %s', err)
+            sys.exit(1)
+        except TypeError:
+            pass
+        except FileNotFoundError:
+            pass
 
     mqttusername = None
     mqttpassword = None
@@ -268,8 +291,9 @@ def main():  # noqa: C901  # pylint: disable=too-many-branches,too-many-statemen
     try:
         while True:
             try:
-                weConnect = weconnect.WeConnect(username=username, password=password, updateAfterLogin=False, updateCapabilities=mqttCLient.updateCapabilities,
-                                                updatePictures=mqttCLient.updatePictures, maxAgePictures=args.pictureCache, selective=mqttCLient.selective,
+                weConnect = weconnect.WeConnect(username=username, password=password, spin=spin, updateAfterLogin=False,
+                                                updateCapabilities=mqttCLient.updateCapabilities, updatePictures=mqttCLient.updatePictures,
+                                                maxAgePictures=args.pictureCache, selective=mqttCLient.selective,
                                                 forceReloginAfter=21600)
                 mqttCLient.connectWeConnect(weConnect)
                 break