+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 - No unreleased changes so far
 
+## [0.39.0] - in testing
+### Added
+- Support for S-PIN by adding --spin 1234 or in .netrc file: "account 1234"
+- Support for locking/unlocking (selected cars only)
+
+### Changed
+- Updated API to 0.47.0
+
 ## [0.38.3] - 2022-08-02
 ### Added
 - Attributes for diesel cars
@@ -648,7 +656,8 @@ Send empty message when topic is disabled
 ## [0.1.0] - 2021-05-27
 Initial release
 
-[unreleased]: https://github.com/tillsteinbach/WeConnect-mqtt/compare/v0.38.3...HEAD
+[unreleased]: https://github.com/tillsteinbach/WeConnect-mqtt/compare/v0.39.0...HEAD
+[0.39.0]: https://github.com/tillsteinbach/WeConnect-mqtt/releases/tag/v0.39.0
 [0.38.3]: https://github.com/tillsteinbach/WeConnect-mqtt/releases/tag/v0.38.3
 [0.38.2]: https://github.com/tillsteinbach/WeConnect-mqtt/releases/tag/v0.38.2
 [0.38.1]: https://github.com/tillsteinbach/WeConnect-mqtt/releases/tag/v0.38.1